# docs/trading_alerts_openapi_compress.yaml
openapi: 3.0.3
info:
  title: Trading Alerts SaaS API
  version: 7.1.0
  description: |
    Commercial SaaS platform for trading alerts and signal generation with real-time chart visualization.
    **Stack:** Next.js 15 + Flask MT5 + Dual Payment (Stripe/dLocal)
    **Auth:** NextAuth.js v4 (Google OAuth + Email/Password, JWT sessions)
    **Tiers:** FREE (5 symbols × 3 timeframes = 15 combos), PRO (15 symbols × 9 timeframes = 135 combos)
    **Payment:** Stripe (international, auto-renewal, 7-day trial), dLocal (8 markets: IN, NG, PK, VN, ID, TH, ZA, TR, manual renewal)
  contact:
    name: API Support
    email: support@tradingalerts.com
  license:
    name: Proprietary
    url: https://tradingalerts.com/license
servers:
  - url: https://api.tradingalerts.com
    description: Production server
  - url: https://staging-api.tradingalerts.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    put:
      tags:
        - User
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                image:
                  type: string
                  format: uri
                  nullable: true
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /api/user/preferences:
    get:
      tags:
        - User
      summary: Get user preferences
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  preferences:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPreference'

    put:
      tags:
        - User
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                key:
                  type: string
                value:
                  type: object
              required:
                - category
                - key
                - value
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/webhooks/stripe:
    post:
      tags:
        - E-commerce
      summary: Stripe webhook handler
      description: Handles subscription lifecycle events (internal endpoint)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe event object
      responses:
        '200':
          description: Event processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean

  # ============================================
  # NOTIFICATIONS ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login (includes tier information)
...
  schemas:
...
    UserTier:
      type: string
      enum:
        - FREE
        - PRO
      description: |
        FREE: 5 symbols × 3 timeframes = 15 combinations
        PRO: 15 symbols × 9 timeframes = 135 combinations
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
          description: Profile picture URL
        tier:
          $ref: '#/components/schemas/UserTier'
        emailVerified:
          type: string
          format: date-time
          nullable: true
          description: |
            Verification timestamp. Email/password users: set after verification. Google OAuth: auto-set. Critical for security.
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
        authMethod:
          type: string
          enum: [credentials, google, both]
          description: credentials=email/password only, google=OAuth only, both=linked
    UserPreference:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        key:
          type: string
        value:
          type: object
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
...